powerd add has_barreljack pref `1`
In R108, power_manager will ignore ACPI AC power supply if `has_barreljack` is
false, and `has_barreljack` is false by default.

This results in some devices charging, but the power manager does not pass the
status to the UI. Since the board `amd64-openfyde*` is for different
devices, power_manager better not ignore it.